TPP-10 qa improvements
