+++ src/main/kotlin/no/nav/pensjon/simulator/tjenestepensjon/pre2025/opptjening/OpptjeningsperiodeService.kt
@@ -4,11 +4,50 @@ import no.nav.pensjon.simulator.tjenestepensjon.pre2025.opptjening.error.Duplica
 import no.nav.pensjon.simulator.tjenestepensjon.pre2025.opptjening.error.MissingOpptjeningsperiodeException
 import no.nav.pensjon.simulator.tjenestepensjon.pre2025.simulering.acl.OpptjeningsperiodeDto
 import no.nav.pensjon.simulator.tjenestepensjon.pre2025.stillingsprosent.acl.Stillingsprosent
-import no.nav.pensjon.simulator.tpregisteret.acl.TpOrdningFullDto
+import no.nav.pensjon.simulator.tpregisteret.TpOrdningFullDto
 import no.nav.tjenestepensjon.simulering.v2.service.OpptjeningsperiodeResponse
+import org.springframework.stereotype.Service
 
-interface OpptjeningsperiodeService {
-    fun getOpptjeningsperiodeListe(tpOrdning: TpOrdningFullDto, stillingsprosentListe: List<Stillingsprosent>): OpptjeningsperiodeResponse
-    @Throws(DuplicateOpptjeningsperiodeEndDateException::class, MissingOpptjeningsperiodeException::class)
-    fun getLatestFromOpptjeningsperiode(map: Map<TpOrdningFullDto, List<OpptjeningsperiodeDto>>): TpOrdningFullDto
+@Service
+class OpptjeningsperiodeService {
+
+    fun getOpptjeningsperiodeListe(tpOrdning: TpOrdningFullDto, stillingsprosentListe: List<Stillingsprosent>) =
+        OpptjeningsperiodeResponse(
+            mapOf(tpOrdning to mapStillingsprosentToOpptjeningsperiodeList(stillingsprosentListe)),
+            emptyList()
+        )
+
+    @Throws(
+            DuplicateOpptjeningsperiodeEndDateException::class,
+            MissingOpptjeningsperiodeException::class
+    )
+    fun getLatestFromOpptjeningsperiode(map: Map<TpOrdningFullDto, List<OpptjeningsperiodeDto>>) =
+            map.flatMap { (key, list) ->
+                list.map { value ->
+                    key to value
+                }
+            }.ifEmpty { throw MissingOpptjeningsperiodeException("Could not find any stillingsprosent") }
+                    .reduce(::getLatest).first
+
+    @Throws(DuplicateOpptjeningsperiodeEndDateException::class)
+    private fun getLatest(latest: Pair<TpOrdningFullDto, OpptjeningsperiodeDto>, other: Pair<TpOrdningFullDto, OpptjeningsperiodeDto>) = when {
+        latest.second.datoTom?.equals(other.second.datoTom) == true -> throw DuplicateOpptjeningsperiodeEndDateException("Could not decide latest stillingprosent due to multiple stillingsprosent having the same end date")
+        other.second.datoTom == null -> other
+        latest.second.datoTom == null || latest.second.datoTom!! > other.second.datoTom -> latest
+        latest.second.datoTom!! < other.second.datoTom -> other
+        else -> latest
+    }
+
+    fun mapStillingsprosentToOpptjeningsperiodeList(stillingsprosentList: List<Stillingsprosent>): List<OpptjeningsperiodeDto> {
+        return stillingsprosentList.map {
+            OpptjeningsperiodeDto(
+                    datoFom = it.datoFom,
+                    datoTom = it.datoTom,
+                    stillingsprosent = it.stillingsprosent,
+                    aldersgrense = it.aldersgrense,
+                    faktiskHovedlonn = it.faktiskHovedlonn?.toIntOrNull(),
+                    stillingsuavhengigTilleggslonn = it.stillingsuavhengigTilleggslonn?.toIntOrNull()
+            )
+        }
+    }
 }
